Restructure benchmark commands
