+++ utils/frame/benchmarking-cli/src/shared/stats.rs
@@ -15,46 +15,38 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-//! Calculates statistics and fills out the `weight.hbs` template.
+//! Handles statistics that were generated from benchmarking results and
+//! that can be used to fill out weight templates.
 
 use sc_cli::Result;
-use sc_service::Configuration;
 
-use log::info;
 use serde::Serialize;
-use std::{fmt, fs, path::PathBuf, result, str::FromStr, time::Duration};
-
-/// Raw output of a Storage benchmark.
-#[derive(Debug, Default, Clone, Serialize)]
-pub(crate) struct BenchRecord {
-	/// Multi-Map of value sizes and the time that it took to access them.
-	ns_per_size: Vec<(u64, u64)>,
-}
+use std::{fmt, result, str::FromStr};
 
 /// Various statistics that help to gauge the quality of the produced weights.
 /// Will be written to the weight file and printed to console.
 #[derive(Serialize, Default, Clone)]
-pub(crate) struct Stats {
+pub struct Stats {
 	/// Sum of all values.
-	pub(crate) sum: u64,
+	pub sum: u64,
 	/// Minimal observed value.
-	pub(crate) min: u64,
+	pub min: u64,
 	/// Maximal observed value.
-	pub(crate) max: u64,
+	pub max: u64,
 
 	/// Average of all values.
-	pub(crate) avg: u64,
+	pub avg: u64,
 	/// Median of all values.
-	pub(crate) median: u64,
+	pub median: u64,
 	/// Standard derivation of all values.
-	pub(crate) stddev: f64,
+	pub stddev: f64,
 
 	/// 99th percentile. At least 99% of all values are below this threshold.
-	pub(crate) p99: u64,
+	pub p99: u64,
 	/// 95th percentile. At least 95% of all values are below this threshold.
-	pub(crate) p95: u64,
+	pub p95: u64,
 	/// 75th percentile. At least 75% of all values are below this threshold.
-	pub(crate) p75: u64,
+	pub p75: u64,
 }
 
 /// Selects a specific field from a [`Stats`] object.
@@ -73,44 +65,6 @@ pub enum StatSelect {
 	P95Percentile,
 	/// Select the 75th percentile.
 	P75Percentile,
-}
-
-impl BenchRecord {
-	/// Appends a new record. Uses safe casts.
-	pub fn append(&mut self, size: usize, d: Duration) -> Result<()> {
-		let size: u64 = size.try_into().map_err(|e| format!("Size overflow u64: {}", e))?;
-		let ns: u64 = d
-			.as_nanos()
-			.try_into()
-			.map_err(|e| format!("Nanoseconds overflow u64: {}", e))?;
-		self.ns_per_size.push((size, ns));
-		Ok(())
-	}
-
-	/// Returns the statistics for *time* and *value size*.
-	pub(crate) fn calculate_stats(self) -> Result<(Stats, Stats)> {
-		let (size, time): (Vec<_>, Vec<_>) = self.ns_per_size.into_iter().unzip();
-		let size = Stats::new(&size)?;
-		let time = Stats::new(&time)?;
-		Ok((time, size)) // The swap of time/size here is intentional.
-	}
-
-	/// Unless a path is specified, saves the raw results in a json file in the current directory.
-	/// Prefixes it with the DB name and suffixed with `path_suffix`.
-	pub fn save_json(&self, cfg: &Configuration, out_path: &PathBuf, suffix: &str) -> Result<()> {
-		let mut path = PathBuf::from(out_path);
-		if path.is_dir() || path.as_os_str().is_empty() {
-			path.push(&format!("{}_{}", cfg.database, suffix).to_lowercase());
-			path.set_extension("json");
-		}
-
-		let json = serde_json::to_string_pretty(&self)
-			.map_err(|e| format!("Serializing as JSON: {:?}", e))?;
-
-		fs::write(&path, json)?;
-		info!("Raw data written to {:?}", fs::canonicalize(&path)?);
-		Ok(())
-	}
 }
 
 impl Stats {
@@ -137,7 +91,7 @@ impl Stats {
 	}
 
 	/// Returns the selected stat.
-	pub(crate) fn select(&self, s: StatSelect) -> u64 {
+	pub fn select(&self, s: StatSelect) -> u64 {
 		match s {
 			StatSelect::Maximum => self.max,
 			StatSelect::Average => self.avg,
